+++ Java/Clase1/ar/com/system2023/mundo/pc/DispositivoEntrada.java
@@ -1,13 +1,14 @@
-package ar.com.system2023.mundopc;
+
+package ar.com.system2023.mundo.pc;
 
 public class DispositivoEntrada {
-
     private String tipoEntrada;
     private String marca;
-
-    public DispositivoEntrada(String topoEntrada, String marca){
+            
+    public DispositivoEntrada(String tipoEntrada, String marca){
         this.tipoEntrada = tipoEntrada;
         this.marca = marca;
+        
     }
 
     public String getTipoEntrada() {
@@ -24,11 +25,12 @@ public class DispositivoEntrada {
 
     public void setMarca(String marca) {
         this.marca = marca;
+        
+        
     }
 
     @Override
     public String toString() {
-        return "DispositivoEntrada [tipoEntrada = " + tipoEntrada + ", marca = " + marca + "]";
+        return "DispositivoEntrada{" + "tipoEntrada=" + tipoEntrada + ", marca=" + marca + '}';
     }
-    
 }
