--- Java/Leccion01/ar/com/system2023/mundopc/mundopc.uxf
@@ -1,72 +1,0 @@
-<diagram program="umletino" version="15.1"><zoom_level>10</zoom_level><element><id>UMLClass</id><coordinates><x>60</x><y>60</y><w>280</w><h>190</h></coordinates><panel_attributes>&lt;&lt;Computadora&gt;&gt;
---
-+contadorComputadoras: int &lt;&lt;static&gt;&gt;
--idComputadora: int
--nombre: String
--monitor: Monitor
--teclado: Teclado
--raton: Raton
---
-+ __str__()
-+ agregar los metodos get/set de cada atributo
---
-Responsibilities
--- Crear objetos de tipo Computadora</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>330</x><y>110</y><w>200</w><h>40</h></coordinates><panel_attributes>lt=&lt;&lt;&lt;&lt;-
-&lt;&lt;agregar&gt;&gt;</panel_attributes><additional_attributes>180;20;10;20</additional_attributes></element><element><id>UMLClass</id><coordinates><x>510</x><y>60</y><w>320</w><h>170</h></coordinates><panel_attributes>&lt;&lt;Orden&gt;&gt;
---
-+ contadorOrdenes: int &lt;&lt;static&gt;&gt;
--idOrden: int
--computadoras: List &lt;&lt;computadora&gt;&gt;
---
-+agregarComputadora(computadora)
-+__str__()
-+(agregar los metodos get/set de cada atributo)
---
-Responsibilities
--- Almacenar en una lista opjetos de tipo Computadora</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>170</x><y>240</y><w>30</w><h>190</h></coordinates><panel_attributes>lt=&lt;&lt;&lt;&lt;-</panel_attributes><additional_attributes>10;10;10;170</additional_attributes></element><element><id>UMLPackage</id><coordinates><x>50</x><y>30</y><w>850</w><h>550</h></coordinates><panel_attributes>																														mundopc
---
-</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLClass</id><coordinates><x>60</x><y>410</y><w>280</w><h>160</h></coordinates><panel_attributes>&lt;&lt;Monitor&gt;&gt;
---
-+contadorMonitores: int &lt;&lt;static&gt;&gt;
--idMonitor: int
--marca: String
--tamaño: String
---
-+ __str__()
-+ agregar los metodos get/set de cada atributo
---
-Responsibilities
--- Crear objetos de tipo Monitor</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>170</x><y>280</y><w>250</w><h>40</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>10;20;230;20</additional_attributes></element><element><id>Relation</id><coordinates><x>390</x><y>290</y><w>30</w><h>290</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>10;10;10;270</additional_attributes></element><element><id>Relation</id><coordinates><x>390</x><y>540</y><w>390</w><h>40</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>370;20;10;20</additional_attributes></element><element><id>UMLClass</id><coordinates><x>420</x><y>420</y><w>210</w><h>120</h></coordinates><panel_attributes>&lt;&lt;Raton&gt;&gt;
---
-+contadorRatones: int &lt;&lt;static&gt;&gt;
--idRaton: int
---
-+ __str__()
---
-Responsibilities
--- Crear objetos de tipo Raton</panel_attributes><additional_attributes></additional_attributes></element><element><id>UMLClass</id><coordinates><x>650</x><y>420</y><w>210</w><h>120</h></coordinates><panel_attributes>&lt;&lt;Teclado&gt;&gt;
---
-+contadorTeclados: int &lt;&lt;static&gt;&gt;
--idTeclado: int
---
-+ __str__()
---
-Responsibilities
--- Crear objetos de tipo Teclado</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>520</x><y>530</y><w>30</w><h>50</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>10;10;10;30</additional_attributes></element><element><id>Relation</id><coordinates><x>750</x><y>530</y><w>30</w><h>50</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>10;10;10;30</additional_attributes></element><element><id>Relation</id><coordinates><x>520</x><y>390</y><w>30</w><h>50</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>10;10;10;30</additional_attributes></element><element><id>Relation</id><coordinates><x>750</x><y>390</y><w>30</w><h>50</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>10;10;10;30</additional_attributes></element><element><id>Relation</id><coordinates><x>520</x><y>380</y><w>260</w><h>40</h></coordinates><panel_attributes>lt=-
-</panel_attributes><additional_attributes>240;20;10;20</additional_attributes></element><element><id>UMLClass</id><coordinates><x>510</x><y>240</y><w>250</w><h>130</h></coordinates><panel_attributes>&lt;&lt;DispositivoEntrada&gt;&gt;
---
-#tipoEntrada: String
-#marca: String
---
-+ __str__()
-+agregarComputadora(computadora)
---
-Responsibilities
--- Crear objetos de tipo DispositivoEntrada</panel_attributes><additional_attributes></additional_attributes></element><element><id>Relation</id><coordinates><x>630</x><y>360</y><w>30</w><h>60</h></coordinates><panel_attributes>lt=&lt;&lt;&lt;-</panel_attributes><additional_attributes>10;10;10;40</additional_attributes></element></diagram>
