--- Java/MundoPc/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=130e46c7
-build.xml.script.CRC32=7f4862c8
-build.xml.stylesheet.CRC32=f85dc8f2@1.96.0.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=130e46c7
-nbproject/build-impl.xml.script.CRC32=afc2a25c
-nbproject/build-impl.xml.stylesheet.CRC32=f89f7d21@1.96.0.48
