--- Java/MundoPc/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\User\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
