--- Java/MundoPc/nbproject/private/private.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/ar/com/system2024/mundo/pc/Teclado.java</file>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/ar/com/system2024/mundo/pc/Monitor.java</file>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/ar/com/system2024/mundo/pc/Raton.java</file>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/mundopc/MundoPC.java</file>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/ar/com/system2024/mundo/pc/Computadora.java</file>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/ar/com/system2024/mundo/pc/Orden.java</file>
-            <file>file:/C:/Users/User/Tecnicatura3S/Java/MundoPc/src/ar/com/system2024/mundo/pc/DispositivoEntrada.java</file>
-        </group>
-    </open-files>
-</project-private>
